Rename stereotype toggle
