+++ src/main/components/controls/icon/stereotype-off.tsx
@@ -3,11 +3,11 @@ import { Icon } from './icon';
 
 type Props = SVGAttributes<SVGSVGElement>;
 
-export const VisibilityOffIcon = (props: Props) => (
+export const StereotypeOffIcon = (props: Props) => (
   <Icon viewBox="0 0 128 128" {...props}>
     <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
+      fillRule="evenodd"
+      clipRule="evenodd"
       d="M82.1654 32.6199C82.0806 32.6106 81.9949 32.6059 81.9085 32.6059H74.1387C73.2767 32.6059 72.4831 33.0754 72.068 33.8309C71.653 34.5864 71.6824 35.508 72.1447 36.2356L74.6334 40.1518L82.1654 32.6199ZM42.8297 71.9556L37.7981 64.0579L55.5222 36.2378C55.9856 35.5104 56.0158 34.5883 55.601 33.8321C55.1861 33.0759 54.3922 32.6059 53.5297 32.6059H45.7599C44.9738 32.6059 44.2392 32.9969 43.8002 33.6489L24.6599 62.0751C24.3973 62.465 24.257 62.9245 24.257 63.3946V64.7212C24.257 65.1913 24.3973 65.6507 24.6599 66.0407L34.3513 80.4339L42.8297 71.9556ZM47.3686 95.5099L53.7626 89.116L55.5222 91.8779C55.9856 92.6054 56.0158 93.5275 55.601 94.2837C55.1861 95.0399 54.3922 95.5099 53.5297 95.5099H47.3686ZM28.0126 86.7726L19.9011 94.8841C19.7674 94.7608 19.647 94.6212 19.5431 94.4669L0.402837 66.0407C0.140261 65.6507 0 65.1913 0 64.7212V63.3946C0 62.9245 0.140261 62.465 0.402837 62.0751L19.5431 33.6489C19.9822 32.9969 20.7168 32.6059 21.5028 32.6059H29.2727C30.1352 32.6059 30.9291 33.0759 31.344 33.8321C31.7588 34.5883 31.7286 35.5104 31.2652 36.2378L13.5411 64.0579L28.0126 86.7726ZM85.5494 57.3292L94.0734 48.8051L103.009 62.0751C103.271 62.465 103.411 62.9245 103.411 63.3946V64.7212C103.411 65.1913 103.271 65.6507 103.009 66.0407L83.8682 94.4669C83.4292 95.1189 82.6946 95.5099 81.9085 95.5099H74.1387C73.2767 95.5099 72.4831 95.0404 72.068 94.2849C71.653 93.5294 71.6824 92.6078 72.1447 91.8802L89.8253 64.0579L85.5494 57.3292ZM100.584 42.2947L114.414 64.0579L96.7334 91.8802C96.271 92.6078 96.2416 93.5294 96.6567 94.2849C97.0717 95.0404 97.8653 95.5099 98.7273 95.5099H106.497C107.283 95.5099 108.018 95.1189 108.457 94.4669L127.597 66.0407C127.86 65.6507 128 65.1913 128 64.7212V63.3946C128 62.9245 127.86 62.465 127.597 62.0751L108.768 34.1107L100.584 42.2947Z"
       fill="black"
     />
