+++ src/main/packages/uml-component-diagram/uml-component-subsystem/uml-component-subsystem-update.tsx
@@ -10,7 +10,7 @@ import { styled } from '../../../components/theme/styles';
 import { UMLElementRepository } from '../../../services/uml-element/uml-element-repository';
 import { AsyncDispatch } from '../../../utils/actions/actions';
 import { IUMLSubsystem, UMLSubsystem } from './uml-component-subsystem';
-import { ToggleButton } from '../../../components/controls/toggle-button/toggle-button';
+import { ToggleButton } from '../../../components/controls/stereotype-toggle/stereotype-toggle';
 
 const Flex = styled.div`
   display: flex;
