+++ src/main/packages/uml-deployment-diagram/uml-deployment-node/uml-deployment-node-update.tsx
@@ -11,7 +11,7 @@ import { styled } from '../../../components/theme/styles';
 import { UMLElementRepository } from '../../../services/uml-element/uml-element-repository';
 import { AsyncDispatch } from '../../../utils/actions/actions';
 import { IUMLDeploymentNode, UMLDeploymentNode } from './uml-deployment-node';
-import { ToggleButton } from '../../../components/controls/toggle-button/toggle-button';
+import { ToggleButton } from '../../../components/controls/stereotype-toggle/stereotype-toggle';
 
 const Flex = styled.div`
   display: flex;
